+++ changelog.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## Unreleased
+
+### Added
+
+- Add prompt to choose between official or custom plugins via `install` command.
+- Externalize official plugins list in leto-modelizer.
+
 ## [0.2.1] - 2022/11/24
 
 ### Fixed